Split MMDetectionTask and OTXDetectionTask
